+++ tests/testthat/test_add_cuts_portfolio.R
@@ -21,7 +21,6 @@ test_that("compile", {
 
 test_that("solve (number_solutions within limit, single zone, Rsymphony)", {
   skip_on_cran()
-  skip_on_ci()
   skip_if_not_installed("Rsymphony")
   # create data
   cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 4))
@@ -46,7 +45,6 @@ test_that("solve (number_solutions within limit, single zone, Rsymphony)", {
 
 test_that("solve (number_solutions within limit, single zone, gurobi)", {
   skip_on_cran()
-  skip_on_ci()
   skip_if_not_installed("gurobi")
   # create data
   cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 4))
@@ -70,7 +68,6 @@ test_that("solve (number_solutions within limit, single zone, gurobi)", {
 
 test_that("solve (number_solutions within limit, multiple zones, Rsymphony)", {
   skip_on_cran()
-  skip_on_ci()
   skip_if_not_installed("Rsymphony")
   # create problem
   data(sim_pu_zones_stack, sim_features_zones)
@@ -103,7 +100,6 @@ test_that("solve (number_solutions within limit, multiple zones, Rsymphony)", {
 
 test_that("solve (number_solutions within limit, multiple zones, gurobi)", {
   skip_on_cran()
-  skip_on_ci()
   skip_if_not_installed("Rsymphony")
   # create problem
   data(sim_pu_zones_stack, sim_features_zones)
@@ -136,7 +132,6 @@ test_that("solve (number_solutions within limit, multiple zones, gurobi)", {
 
 test_that("solve (number_solutions outside limit, Rsymphony)", {
   skip_on_cran()
-  skip_on_ci()
   skip_if_not_installed("Rsymphony")
   # create data
   cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 4))
@@ -169,7 +164,6 @@ test_that("solve (number_solutions outside limit, Rsymphony)", {
 
 test_that("solve (number_solutions outside limit, gurobi)", {
   skip_on_cran()
-  skip_on_ci()
   skip_if_not_installed("gurobi")
   # create data
   cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 4))