+++ tests/testthat/test_add_feature_contiguity_constraints.R
@@ -115,8 +115,7 @@ test_that("compile (single zone)", {
 
 test_that("solve (single zone)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create data
   spp1_habitat <- raster::raster(matrix(c(
     5, 0, 5,
@@ -350,8 +349,7 @@ test_that("compile (multiple zones)", {
 
 test_that("solve (multiple zones)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create data
   spp1_z1 <- raster::raster(matrix(c(
     5, 0, 5,