+++ tests/testthat/test_add_locked_in_constraints.R
@@ -24,8 +24,7 @@ test_that("integer (compile, single zone)", {
 
 test_that("integer (solve, single zone)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create problem
   data(sim_pu_raster, sim_features)
   suppressWarnings({
@@ -69,8 +68,7 @@ test_that("logical (compile, single zone)", {
 
 test_that("logical (solve, single zone)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create problem
   data(sim_pu_raster, sim_features)
   suppressWarnings({
@@ -123,8 +121,7 @@ test_that("matrix (compile, multiple zones)", {
 
 test_that("matrix (solve, multiple zones)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create problem
   data(sim_pu_zones_stack, sim_features_zones)
   status <- matrix(FALSE, nrow = raster::ncell(sim_pu_zones_stack),
@@ -196,8 +193,7 @@ test_that("character (compile, Spatial, single zone)", {
 
 test_that("character (solve, Spatial, single zone)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create problem
   data(sim_pu_polygons, sim_features)
   s <- problem(sim_pu_polygons, sim_features, "cost") %>%
@@ -264,8 +260,7 @@ test_that("character (compile, Spatial, multiple zones)", {
 
 test_that("character (solve, Spatial, multiple zones)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create problem
   data(sim_pu_zones_polygons, sim_features_zones)
   targets <- matrix(FALSE, nrow = number_of_features(sim_features_zones),
@@ -291,8 +286,7 @@ test_that("character (solve, Spatial, multiple zones)", {
 
 test_that("character (solve, single zone, proportion decisions)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create problem
   data(sim_pu_polygons, sim_features)
   s <- problem(sim_pu_polygons, sim_features, "cost") %>%
@@ -308,7 +302,7 @@ test_that("character (solve, single zone, proportion decisions)", {
 
 test_that("character (solve, multiple zones, proportion decisions)", {
   skip_on_cran()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   data(sim_pu_zones_polygons, sim_features_zones)
   targets <- matrix(FALSE, nrow = number_of_features(sim_features_zones),
                     ncol = number_of_zones(sim_features_zones))
@@ -389,8 +383,7 @@ test_that("raster (compile, single zone)", {
 
 test_that("raster (solve, single zone)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create problem
   data(sim_pu_raster, sim_locked_in_raster, sim_features)
   s <- problem(sim_pu_raster, sim_features) %>%
@@ -447,8 +440,7 @@ test_that("raster (compile, multiple zones)", {
 
 test_that("raster (solve, multiple zones)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create problem
   data(sim_pu_zones_stack, sim_features_zones)
   targets <- matrix(FALSE, nrow = number_of_features(sim_features_zones),
@@ -521,8 +513,7 @@ test_that("spatial (compile, single zone)", {
 
 test_that("spatial (solve, single zone)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create problem
   data(sim_pu_polygons, sim_features)
   s <- problem(sim_pu_polygons, sim_features, "cost") %>%