+++ tests/testthat/test_add_min_largest_shortfall_objective.R
@@ -43,8 +43,7 @@ test_that("compile (compressed formulation, single zone)", {
 
 test_that("solution (compressed formulation, single zone)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create data
   budget <- 4.23
   cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 4))
@@ -131,8 +130,7 @@ test_that("compile (expanded formulation, single zone)", {
 
 test_that("solution (expanded formulation, single zone)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create data
   budget <- 4.23
   cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 4))
@@ -154,6 +152,7 @@ test_that("solution (expanded formulation, single zone)", {
 })
 
 test_that("invalid inputs (single zone)", {
+  data(sim_pu_raster, sim_features, sim_phylogeny)
   # check that invalid arguments result in errors
   expect_error({
     problem(sim_pu_raster, sim_features) %>%
@@ -174,6 +173,12 @@ test_that("invalid inputs (single zone)", {
     problem(sim_pu_raster, sim_features) %>%
       add_min_largest_shortfall_objective(budget = Inf) %>%
       add_absolute_targets(targ)
+  })
+  # check that no targets results in error
+  expect_error({
+    problem(sim_pu_raster, sim_pu_raster) %>%
+    add_min_largest_shortfall_objective(budget = 5) %>%
+    compile()
   })
 })
 
@@ -247,8 +252,7 @@ test_that("compile (compressed formulation, multiple zones, scalar budget)", {
 
 test_that("solve (compressed formulation, multiple zones, scalar budget)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # make and solve problem
   budget <- 7
   locked_out <- matrix(FALSE, ncol = 2, nrow = 5)
@@ -356,8 +360,7 @@ test_that("compile (compressed formulation, multiple zones, vector budget)", {
 
 test_that("solve (compressed formulation, multiple zones, vector budget)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # make and solve problem
   budget <- c(6, 1)
   locked_out <- matrix(FALSE, ncol = 2, nrow = 5)
@@ -489,8 +492,7 @@ test_that("compile (expanded formulation, multiple zones, vector budget)", {
 
 test_that("solve (expanded formulation, multiple zones, vector budget)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # make and solve problem
   budget <- c(6, 1)
   locked_out <- matrix(FALSE, ncol = 2, nrow = 5)
@@ -526,6 +528,7 @@ test_that("solve (expanded formulation, multiple zones, vector budget)", {
 
 test_that("invalid inputs (multiple zones)", {
   data(sim_pu_zones_stack, sim_features_zones)
+  # check that no targets results in error
   expect_error({
     problem(sim_pu_zones_stack, sim_features_zones) %>%
     add_min_largest_shortfall_objective(budget = c(1, -5, 1))
@@ -545,5 +548,11 @@ test_that("invalid inputs (multiple zones)", {
   expect_error({
     problem(sim_pu_zones_stack, sim_features_zones) %>%
     add_min_largest_shortfall_objective(budget = c(1, Inf, 9))
+  })
+  # check that no targets results in error
+  expect_error({
+    problem(sim_pu_zones_stack, sim_features_zones) %>%
+    add_min_largest_shortfall_objective(budget = c(5, 5, 5)) %>%
+    compile()
   })
 })