+++ tests/testthat/test_add_shuffle_portfolio.R
@@ -19,8 +19,7 @@ test_that("compile", {
 
 test_that("solve (RasterLayer, single zone)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create data
   cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 4))
   features <- raster::stack(raster::raster(matrix(c(2, 1, 1, 0), ncol = 4)),
@@ -46,8 +45,7 @@ test_that("solve (RasterLayer, single zone)", {
 
 test_that("solve (RasterStack, multiple zones)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create data
   data(sim_pu_zones_stack, sim_features_zones)
   # create problem
@@ -74,8 +72,7 @@ test_that("solve (RasterStack, multiple zones)", {
 
 test_that("solve (SpatialPolygonsDataFrame, single zone)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create data
   data(sim_pu_polygons, sim_features)
   # create problem
@@ -98,8 +95,7 @@ test_that("solve (SpatialPolygonsDataFrame, single zone)", {
 
 test_that("solve (SpatialPolygonsDataFrame, multiple zones)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # make data
   data(sim_pu_zones_polygons, sim_features_zones)
   # solve problem
@@ -130,8 +126,7 @@ test_that("solve (SpatialPolygonsDataFrame, multiple zones)", {
 
 test_that("solve (numeric, single zone)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create data
   data(sim_pu_polygons, sim_features)
   costs <- sim_pu_polygons$cost
@@ -161,8 +156,7 @@ test_that("solve (numeric, single zone)", {
 
 test_that("solve (matrix, multiple zones)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create data
   costs <- data.frame(id = seq_len(7),
                       cost_1 = c(1,  2,  NA, 3, 100, 100, NA),
@@ -205,8 +199,7 @@ test_that("solve (matrix, multiple zones)", {
 
 test_that("solve (no duplicates)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create data
   set.seed(500)
   cost <- raster::raster(matrix(c(1, 1, 0.5, NA), ncol = 4))
@@ -234,8 +227,7 @@ test_that("solve (no duplicates)", {
 
 test_that("solve (parallel processing)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # create data
   set.seed(500)
   cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 4))