+++ tests/testthat/test_add_top_portfolio.R
@@ -20,7 +20,6 @@ test_that("compile", {
 
 test_that("solve (RasterLayer, single zone)", {
   skip_on_cran()
-  skip_on_ci()
   skip_if_not_installed("gurobi")
   # create data
   cost <- raster::raster(matrix(c(1, 2, 2, NA), ncol = 4))
@@ -47,7 +46,6 @@ test_that("solve (RasterLayer, single zone)", {
 
 test_that("solve (RasterStack, multiple zones)", {
   skip_on_cran()
-  skip_on_ci()
   skip_if_not_installed("gurobi")
   # create data
   data(sim_pu_zones_stack, sim_features_zones)
@@ -75,7 +73,6 @@ test_that("solve (RasterStack, multiple zones)", {
 
 test_that("solve (SpatialPolygonsDataFrame, single zone)", {
   skip_on_cran()
-  skip_on_ci()
   skip_if_not_installed("gurobi")
   # create data
   data(sim_pu_polygons, sim_features)
@@ -101,7 +98,6 @@ test_that("solve (SpatialPolygonsDataFrame, single zone)", {
 
 test_that("solve (SpatialPolygonsDataFrame, multiple zones)", {
   skip_on_cran()
-  skip_on_ci()
   skip_if_not_installed("gurobi")
   # make data
   data(sim_pu_zones_polygons, sim_features_zones)
@@ -137,7 +133,6 @@ test_that("solve (SpatialPolygonsDataFrame, multiple zones)", {
 
 test_that("solve (numeric, single zone)", {
   skip_on_cran()
-  skip_on_ci()
   skip_if_not_installed("gurobi")
   # create data
   data(sim_pu_polygons, sim_features)
@@ -169,7 +164,6 @@ test_that("solve (numeric, single zone)", {
 
 test_that("solve (matrix, multiple zones)", {
   skip_on_cran()
-  skip_on_ci()
   skip_if_not_installed("gurobi")
   # create data
   costs <- data.frame(id = seq_len(7),