+++ tests/testthat/test_marxan_problem.R
@@ -39,8 +39,7 @@ test_that("character (compile)", {
 
 test_that("character (solve)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # make problem
   path <- system.file("extdata/input.dat", package = "prioritizr")
   p <- marxan_problem(path) %>%
@@ -55,8 +54,7 @@ test_that("character (solve)", {
 
 test_that("character (solve, absolute INPUTDIR path)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # set up input.dat with absolute file paths
   path <- file.path(tempfile(fileext = ".dat"))
   f <- readLines(system.file("extdata/input.dat", package = "prioritizr"))
@@ -76,8 +74,7 @@ test_that("character (solve, absolute INPUTDIR path)", {
 
 test_that("character (solve, absolute file paths)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # set up input.dat with absolute file paths
   path <- file.path(tempfile(fileext = ".dat"))
   f <- readLines(system.file("extdata/input.dat", package = "prioritizr"))
@@ -175,8 +172,7 @@ test_that("data.frame (compile, no boundary penalties)", {
 
 test_that("data.frame (solve, boundary penalties)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # make problem
   path <- system.file("extdata/input.dat", package = "prioritizr")
   wd <- system.file("extdata/input", package = "prioritizr")
@@ -198,8 +194,7 @@ test_that("data.frame (solve, boundary penalties)", {
 
 test_that("data.frame (solve, no boundary penalties)", {
   skip_on_cran()
-  skip_on_ci()
-  skip_if_not(any_solvers_installed())
+  skip_if_no_fast_solvers_installed()
   # make problem
   path <- system.file("extdata/input.dat", package = "prioritizr")
   wd <- system.file("extdata/input", package = "prioritizr")